+++ HISTORY.rst
@@ -18,6 +18,7 @@ PlatformIO 4.0
 * Override default source and include directories for a library via `library.json <http://docs.platformio.org/page/librarymanager/config.html>`__ manifest using ``includeDir`` and ``srcDir`` fields
 * Added support for the latest Python "Click" package (CLI Builder)
   (`issue #349 <https://github.com/platformio/platformio-core/issues/349>`_)
+* Deprecated ``--only-check`` CLI option for "update" sub-commands, please use ``--dry-run`` instead
 
 PlatformIO 3.0
 --------------